get pile validation set and small edits to make code run on zephyr model

+++ tuned_lens/scripts/train_loop.py
@@ -17,6 +17,7 @@ from torch.optim.lr_scheduler import LambdaLR
 from torchdata.dataloader2 import DataLoader2
 from tqdm.auto import trange
 from transformers import PreTrainedModel
+import torch
 
 import tuned_lens.scripts.ingredients as ing
 from tuned_lens import TunedLens
@@ -413,7 +414,7 @@ class Train:
                 # We use bfloat16 because it has a larger dynamic range than float16
                 # and it seems to remove the need for doing grad scaling, which is very
                 # annoying to set up in the context of multiple backward passes.
-                with th.autocast(self.dist.device.type, dtype=th.bfloat16):
+                with th.autocast(self.dist.device.type, dtype=torch.float16):
                     logits = shift_preds(state.lens(h, idx=i), shift)
 
                     if self.loss == LossChoice.CE:
